requirements.txt
numpy
pandas
os
openai
flask
flask_socketio
boto3